Struct net_adds::Ipv6AddrRange

pub struct Ipv6AddrRange { /* fields omitted */ }

An IPv6 address range.

See crate::IpAddrRange for a type encompassing both IPv4 and IPv6 range.

The size of an Ipv6AddrRange struct may vary depending on the target operating system.

Textual representation

Ipv6AddrRange provides a FromStr implementation. The two parts are represented by Ipv6Addr and are separated by ‘-’.

Examples

use std::net::Ipv6Addr;

use net_adds::Ipv6AddrRange;

let a = Ipv6Addr::from(0x1);
let b = Ipv6Addr::from(0xA);
let range = Ipv6AddrRange::new(a, b);

assert_eq!(Ok(range), "::1-::A".parse());

Implementations

impl Ipv6AddrRange

pub fn new(start: Ipv6Addr, end: Ipv6Addr) -> Ipv6AddrRange

Returns an Ipv6AddrRange.

pub fn start(self) -> Ipv6Addr

Returns the first ip.

pub fn end(self) -> Ipv6Addr

Returns the last ip.

pub fn all(&self) -> Vec<Ipv6Addr>

Returns all ips included in the range.

Examples:

use std::net::Ipv6Addr;

use net_adds::Ipv6AddrRange;

let a = Ipv6Addr::from(0x1);
let b = Ipv6Addr::from(0x2);
let c = Ipv6Addr::from(0x3);

assert_eq!(Ipv6AddrRange::new(a, c).all(), vec![a, b, c]);
assert_eq!(Ipv6AddrRange::new(c, a).all(), vec![c, b, a]);
assert_eq!(Ipv6AddrRange::new(a, a).all(), vec![a]);

pub fn size(&self) -> u128

Returns the number of ip’s included in the range.

Examples:

use std::net::Ipv6Addr;

use net_adds::Ipv6AddrRange;

let a = Ipv6Addr::from(0x0);
let b = Ipv6Addr::from(0xA);

let range = Ipv6AddrRange::new(a, b);
assert_eq!(range.size(), 11);

let range = Ipv6AddrRange::new(b, a);
assert_eq!(range.size(), 11);

pub fn has(&self, ip: Ipv6Addr) -> bool

Returns true if the ip argument is included in the range, else returns false.

Examples:

use std::net::Ipv6Addr;

use net_adds::Ipv6AddrRange;

let a = Ipv6Addr::from(0x0);
let b = Ipv6Addr::from(0xA);

let range = Ipv6AddrRange::new(a, b);

assert!(range.has(Ipv6Addr::from(0x0)));
assert!(range.has(Ipv6Addr::from(0x2)));
assert!(range.has(Ipv6Addr::from(0xA)));

assert!(!range.has(Ipv6Addr::from(0xFFFF)));

let range = Ipv6AddrRange::new(b, a);

assert!(range.has(Ipv6Addr::from(0x0)));
assert!(range.has(Ipv6Addr::from(0x2)));
assert!(range.has(Ipv6Addr::from(0xA)));

assert!(!range.has(Ipv6Addr::from(0xFFFF)));