Trait net_adds::IntoSmartIterator

pub trait IntoSmartIterator {
    type Item;
    type IntoSmartIter: Iterator;
    fn into_smart_iter(self) -> Self::IntoSmartIter;
}

Conversion into a smart Iterator.

A smart iterator is an iterator that does not store all the items in memory. It’s usefull for long series.

By implementing IntoSmartIterator for a type, you define how it will be converted to an smart iterator.

This is common for types which describe a followed collection of some kind.

Associated Types

type Item

The type of the elements being iterated over.

type IntoSmartIter: Iterator

Which kind of iterator are we turning this into?

Required methods

fn into_smart_iter(self) -> Self::IntoSmartIter

Creates an iterator from the current struct.